requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
pillow==10.2.0
psycopg==3.1.18
psycopg2-binary==2.9.9
